+++ example/app/factories.py
@@ -10,12 +10,12 @@ DEFAULT_PASSWORD = "Test111!"  # noqa: S105
 class UserFactory(factory.django.DjangoModelFactory[models.User]):
     """Factory to generate test User instance."""
 
-    email = factory.LazyAttribute(
+    email = factory.declarations.LazyAttribute(
         lambda obj: f"{uuid.uuid4()}@saritasa-s3-tools.com",
     )
-    username = factory.Faker("user_name")
-    first_name = factory.Faker("first_name")
-    last_name = factory.Faker("last_name")
+    username = factory.faker.Faker("user_name")
+    first_name = factory.faker.Faker("first_name")
+    last_name = factory.faker.Faker("last_name")
     password = factory.django.Password(password=DEFAULT_PASSWORD)
 
     class Meta:
@@ -49,3 +49,12 @@ class ModelWithFilesFactory(
 
     class Meta:
         model = models.ModelWithFiles
+
+
+class ModelWithFilesInvalidTypesFactory(ModelWithFilesFactory):
+    """Factory to generate test ModelWithFiles instance with invalid types."""
+
+    image = factory.django.FileField(
+        filename="file.txt",
+        data="Test",
+    )